Refactor CSS, change some core code relating to data management
* CSS REFACTOR BBY

* back to 3

* Header changes + Logo Additions (#20)

* Added Logos

* Adds Futura

* Updates Header

* Adjusted small-mode header as well.

* New Legendary Items style. (#21)

* Adjusts Cost Counter (#22)

* New Legendary Items style.

* Adjusts Cost Container to be a top level item.

Instead of being inside the body content, which it is seperate from, it
is now solo and doing its own thing like a strong independant cost
container.
(This lets it be positioned based on the page, not the body container.)

* Add about

* Improves Hero Portraits (#23)

* New Legendary Items style.

* Adjusts Cost Container to be a top level item.

Instead of being inside the body content, which it is seperate from, it
is now solo and doing its own thing like a strong independant cost
container.
(This lets it be positioned based on the page, not the body container.)

* Improves the design of the Hero Portraits

* Goddamn it GitHub... Stahp! (Makes the legendary bottom section fully clickable) (#25)

* New Legendary Items style.

* Adjusts Cost Container to be a top level item.

Instead of being inside the body content, which it is seperate from, it
is now solo and doing its own thing like a strong independant cost
container.
(This lets it be positioned based on the page, not the body container.)

* Improves the design of the Hero Portraits

* The entire bottom area of legendary items clickable

* Adds the Legendary clickable stuff to all templates

* Fix per hero,

* Add small css for rooster

* Modified templates to add information. (#26)

Usable information for when other themes are introduced. Moved certain
content to CSS for themability. Base theme from Jayess remains
unaffected visually.

* I think that does it

* concat files

* Loads of shit

* Better safe den sorreh

* Changes header icons to inline SVG (#27)

* there we go